Optimizations
Square grid is now rendered by columns which means that when something changes on the grid, only relevant columns have to be updated in the UI.
Rewritten some update functions in the Grid dao to reduce number of unnecessary updates. (For example Map#map always returns a new reference even if nothing in List changes, eventually triggering UI update).